Remove defunct files and rename 'account-manager.ts' to 'management-sdk.ts' for consistency
